--- Memoria.Launcher/ModManager/ModConflictWindow.xaml
@@ -1,67 +1,0 @@
-﻿<Window x:Class="Memoria.Launcher.ModConflictWindow"
-xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-mc:Ignorable="d"
-Title="{Binding WindowTitle, FallbackValue='Mod Conflicts'}"
-WindowStartupLocation="CenterScreen"
-MinHeight="300" Height="300"
-MinWidth="600" Width="600" >
-
-    <Grid
-    Background="#333" >
-        <Grid.RowDefinitions>
-            <RowDefinition Height="40"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="40"/>
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="300"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-
-        <TextBlock x:Name="OverallVerdict"
-        Margin="10,5,3,0"
-        Grid.Row="0"
-        Grid.Column="0"
-        Grid.ColumnSpan="2"
-        Foreground="#fff"
-        Text="Compatibility info"
-        FontSize="16"
-        FontWeight="SemiBold"
-        TextAlignment="Left"
-        HorizontalAlignment="left"
-        VerticalAlignment="Center"/>
-
-        <ListBox x:Name="ModList"
-        Margin="10,5,3,0"
-        Grid.Row="1"
-        Grid.Column="0"
-        SelectionChanged="OnModSelection"
-        DisplayMemberPath="Label"/>
-
-        <TextBox x:Name="DetailedInfos"
-        Margin="3,5,10,0"
-        Grid.Row="1"
-        Grid.Column="1"
-        Text="Detailed informations"
-        IsReadOnly="True"
-        ScrollViewer.HorizontalScrollBarVisibility="Auto"
-        ScrollViewer.VerticalScrollBarVisibility="Visible"
-        TextWrapping="NoWrap"/>
-
-        <WrapPanel
-        Margin="3,5,3,5"
-        Grid.Row="2"
-        Grid.Column="0"
-        Grid.ColumnSpan="2"
-        HorizontalAlignment="Center"
-        VerticalAlignment="Bottom">
-            <Button x:Name="OkButton"
-            Content="Ok"
-            MinWidth="100"
-            Click="OnButtonClick"/>
-        </WrapPanel>
-    </Grid>
-</Window>
