+++ Memoria.Launcher/Resources/TextBoxStyle.xaml
@@ -1,19 +1,3 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    <!--
-    <Style x:Key="HeadingStyle" TargetType="Border">
-        <Setter Property="Background" Value="#F03a6b77"/>
-        <Setter Property="CornerRadius" Value="10"/>
-        <Setter Property="Padding" Value="10,5"/>
-        <Setter Property="HorizontalAlignment" Value="Stretch"/>
-        <Setter Property="VerticalAlignment" Value="Center"/>
-        <Setter Property="Margin" Value="0,10,0,10"/>
-        <Setter Property="Child">
-            <Setter.Value>
-                <TextBlock TextAlignment="Center" FontSize="16" FontWeight="Normal" Foreground="White"/>
-            </Setter.Value>
-        </Setter>
-    </Style>
-    -->
 </ResourceDictionary>
